requirements.txt
ultralytics>=8.0.0
opencv-python>=4.5.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pyyaml>=5.4.0
tqdm>=4.62.0
